+++ examples/_debug/styles/size-categories-style.html
@@ -66,7 +66,7 @@
 
         const style = carto.viz.sizeCategoriesStyle('type');
 
-        const source = new carto.viz.CARTOSource('roads', {
+        const source = new carto.source.Dataset('roads', {
           mapOptions: {
             bufferSize: {
               mvt: 30
