+++ src/lib/dataviews/category/Category.ts
@@ -1,4 +1,5 @@
-import { Layer, Source } from '@/viz';
+import { Layer } from '@/viz';
+import { Source } from '@/source';
 import { BuiltInFilters } from '@/viz/filters/types';
 import { uuidv4 } from '@/core/utils/uuid';
 import { DataViewLocal } from '../mode/DataViewLocal';
