+++ src/lib/dataviews/mode/DataViewMode.ts
@@ -1,4 +1,5 @@
-import { Source, Layer } from '@/viz';
+import { Layer } from '@/viz';
+import { Source } from '@/source';
 import { WithEvents } from '@/core/mixins/WithEvents';
 import { Filter } from '@/viz/filters/types';
 import { AggregationType } from '@/data/operations/aggregation/aggregation';
