+++ src/lib/viz/__tests__/styles/ColorCategoriesStyle.spec.ts
@@ -1,7 +1,7 @@
 import { Deck } from '@deck.gl/core';
+import { DatasetSource } from '@/source';
 import { colorCategoriesStyle } from '../../style';
 import * as mapsResponse from '../data-mocks/maps.category.json';
-import { CARTOSource } from '../../sources/CARTOSource';
 import { CartoStylingError } from '../../errors/styling-error';
 import { hexToRgb, getColors } from '../../style/helpers/utils';
 import { DEFAULT_PALETTE } from '../../style/helpers/color-categories-style';
@@ -21,13 +21,13 @@ const getMetadata = jest.fn().mockImplementation(() => {
   };
 });
 
-jest.mock('../../sources/CARTOSource', () => ({
-  CARTOSource: jest.fn().mockImplementation(() => ({ getMetadata }))
+jest.mock('../../../source/DatasetSource', () => ({
+  DatasetSource: jest.fn().mockImplementation(() => ({ getMetadata }))
 }));
 
 const styledLayer = {
   getMapInstance: () => ({} as Deck),
-  source: new CARTOSource('table')
+  source: new DatasetSource('table')
 };
 
 describe('ColorCategoriesStyle', () => {
