+++ src/lib/viz/__tests__/styles/ColorContinuousStyle.spec.ts
@@ -1,8 +1,8 @@
 import { Deck } from '@deck.gl/core';
 import { scale as chromaScale } from 'chroma-js';
+import { DatasetSource } from '@/source';
 import { colorContinuousStyle } from '../../style';
 import * as mapsResponse from '../data-mocks/maps.number.json';
-import { CARTOSource } from '../../sources/CARTOSource';
 import { hexToRgb } from '../../style/helpers/utils';
 import { CartoStylingError } from '../../errors/styling-error';
 
@@ -21,13 +21,13 @@ const getMetadata = jest.fn().mockImplementation(() => {
   };
 });
 
-jest.mock('../../sources/CARTOSource', () => ({
-  CARTOSource: jest.fn().mockImplementation(() => ({ getMetadata }))
+jest.mock('../../../source/DatasetSource', () => ({
+  DatasetSource: jest.fn().mockImplementation(() => ({ getMetadata }))
 }));
 
 const styledLayer = {
   getMapInstance: () => ({} as Deck),
-  source: new CARTOSource('table')
+  source: new DatasetSource('table')
 };
 
 describe('ColorContinuousStyle', () => {
