+++ src/lib/viz/__tests__/styles/SizeContinuousStyle.spec.ts
@@ -1,8 +1,8 @@
 import { Deck } from '@deck.gl/core';
+import { DatasetSource } from '@/source';
 import { sizeContinuousStyle } from '../../style';
 import { getDefaultSizeRange } from '../../style/helpers/size-continuous-style';
 import * as mapsResponse from '../data-mocks/maps.number.json';
-import { CARTOSource } from '../../sources/CARTOSource';
 import { CartoStylingError } from '../../errors/styling-error';
 
 const FIELD_NAME = 'pct_higher_ed';
@@ -20,13 +20,13 @@ const getMetadata = jest.fn().mockImplementation(() => {
   };
 });
 
-jest.mock('../../sources/CARTOSource', () => ({
-  CARTOSource: jest.fn().mockImplementation(() => ({ getMetadata }))
+jest.mock('../../../source/DatasetSource', () => ({
+  DatasetSource: jest.fn().mockImplementation(() => ({ getMetadata }))
 }));
 
 const styledLayer = {
   getMapInstance: () => ({} as Deck),
-  source: new CARTOSource('table')
+  source: new DatasetSource('table')
 };
 
 describe('SizeContinuousStyle', () => {
