+++ src/lib/viz/style/helpers/size-bins-style.ts
@@ -1,8 +1,8 @@
+import { NumericFieldStats, GeometryType } from '@/source';
 import { findIndexForBinBuckets, calculateSizeBins } from './utils';
 import { Classifier, ClassificationMethod } from '../../utils/Classifier';
 import { CartoStylingError, stylingErrorTypes } from '../../errors/styling-error';
 import { StyledLayer } from '../layer-style';
-import { NumericFieldStats, GeometryType } from '../../sources/Source';
 import { Style, BasicOptionsStyle, getStyles, getStyleValue } from '..';
 import { sizeRangeValidation } from '../validators';
