+++ src/lib/viz/utils/Classifier.ts
@@ -1,5 +1,5 @@
+import { Stats } from '@/source';
 import { CartoStylingError, stylingErrorTypes } from '../errors/styling-error';
-import { Stats } from '../sources/Source';
 
 export type ClassificationMethod = 'quantiles' | 'stdev' | 'equal';
